requirements.txt
crewai>=1.0.0
crewai-tools>=1.0.0
google-genai>=1.0.0
google-generativeai>=0.8.0
tavily-python>=0.3.0
httpx>=0.27.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
